adding env variables and fixing types
