+++ .env
@@ -1,0 +1,1 @@
+PUBLIC_API_URL='https://post-to-src.onrender.com'
