+++ .gitignore
@@ -11,7 +11,6 @@ node_modules
 Thumbs.db
 
 # Env
-.env
 .env.*
 !.env.example
 !.env.test
