+++ tsconfig.json
@@ -11,8 +11,8 @@
     "strict": true,
     "moduleResolution": "bundler",
     "types": ["vite/client"]
-  },
-  "include": ["glfx-es6.d.ts"]
+  }
+  // "include": ["glfx-es6.d.ts"]
   // Path aliases are handled by https://kit.svelte.dev/docs/configuration#alias
   // except $lib which is handled by https://kit.svelte.dev/docs/configuration#files
   //
